+++ knight-text-adventure/knight-text-adventure/Items/Sword.cs
@@ -1,0 +1,18 @@
+namespace knight_text_adventure.Items;
+{
+    public class Sword : Item
+    {
+        public int Damage { get; set; }
+
+        public Sword(int damage, string name) : base(name)
+        {
+            Name = name;
+            Damage = damage;
+        }
+
+        public override void Use()
+        {
+            Console.WriteLine("Wschhh");
+        }
+    }
+}
